Refactor dialog initialization
Registering dialog code is now extracted into a separate file in the
dialog folder.
