+++ module/burningwheel.ts
@@ -1,8 +1,6 @@
 import { BWActor } from "./actors/BWActor.js";
 import { BWCharacterSheet } from "./actors/sheets/BWCharacterSheet.js";
-import {
-    RegisterItemSheets
-} from "./items/item.js";
+import { RegisterItemSheets } from "./items/item.js";
 
 import { hideChatButtonsIfNotOwner, onChatLogRender } from "./chat.js";
 import { ShadeString, slugify, translateWoundValue } from "./helpers.js";
@@ -10,16 +8,13 @@ import { migrateData } from "./migration/migration.js";
 import { registerSystemSettings } from "./settings.js";
 import { preloadHandlebarsTemplates } from "./templates.js";
 import { NpcSheet } from "./actors/sheets/NpcSheet.js";
-import { DuelOfWitsDialog } from "./dialogs/DuelOfWitsDialog.js";
-import { FightDialog } from "./dialogs/FightDialog.js";
-import { DifficultyDialog } from "./dialogs/DifficultyDialog.js";
-import { RangeAndCoverDialog } from "./dialogs/RangeAndCoverDialog.js";
+
 import { actorConstructor, itemConstructor } from "./factory.js";
 
 import * as constants from "./constants.js";
 import { CreateBurningWheelMacro, RegisterMacros } from "./macros/Macro.js";
 import { BWSettingSheet } from "./actors/sheets/BWSettingSheet.js";
-import { ModifierDialog } from "./dialogs/ModifierDialog.js";
+import * as dialogs from "./dialogs/index.js";
 
 Hooks.once("init", async () => {
     CONFIG.Actor.entityClass = actorConstructor;
@@ -45,86 +40,13 @@ Hooks.once("init", async () => {
     registerSystemSettings();
     preloadHandlebarsTemplates();
     registerHelpers();
-
-    let dowData = {};
-    let fightData = {};
-    let rncData = {};
-    try {
-        dowData = await JSON.parse(game.settings.get(constants.systemName, constants.settings.duelData));
-    } catch (err) {
-        ui.notifications.warn("Error parsing serialized Duel of Wits data");
-        console.log(err);
-    }
-    try {
-        fightData = await JSON.parse(game.settings.get(constants.systemName, constants.settings.fightData));
-    } catch (err) {
-        ui.notifications.warn("Error parsing serialized Fight data");
-        console.log(err);
-    }
-    try {
-        rncData = await JSON.parse(game.settings.get(constants.systemName, constants.settings.rangeData));
-    } catch (err) {
-        ui.notifications.warn("Error parsing serialized Range and Cover data");
-        console.log(err);
-    }
-    
-    game.burningwheel.dow = new DuelOfWitsDialog({
-        title: "Duel of Wits",
-        buttons: {},
-        data: dowData
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    } as any);
-    game.burningwheel.dow.activateSocketListeners();
-
-    game.burningwheel.fight = new FightDialog({
-        title: "Fight!",
-        buttons: {},
-        data: fightData
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    } as any);
-    game.burningwheel.fight.activateSocketListeners();
-
-    game.burningwheel.rangeAndCover = new RangeAndCoverDialog({
-        title: "Range and Cover",
-        buttons: {},
-        data: rncData
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    } as any);
-    game.burningwheel.rangeAndCover.activateSocketListeners();
-
+    dialogs.initializeExtendedTestDialogs();
 });
 
 Hooks.once("ready", async() => {
-    migrateData();
-    game.burningwheel.useGmDifficulty = await game.settings.get(constants.systemName, constants.settings.useGmDifficulty);
-    if (game.burningwheel.useGmDifficulty) {
-        const difficulty = await game.settings.get(constants.systemName, constants.settings.gmDifficulty);
-        const testData = await JSON.parse(game.settings.get(constants.systemName, constants.settings.extendedTestData));
-        game.burningwheel.gmDifficulty = new DifficultyDialog(difficulty, testData);
-        game.burningwheel.gmDifficulty.activateSocketListeners();
-        game.burningwheel.gmDifficulty.render(true);
-    }
-
-    let modData = { mods: undefined, help: undefined };
-    try {
-        modData = await JSON.parse(game.settings.get(constants.systemName, constants.settings.obstacleList));
-    } catch (err) {
-        ui.notifications.warn("Error parsing serialized Modifier data");
-        console.log(err);
-    }
-    game.burningwheel.modifiers = new ModifierDialog(game.burningwheel.useGmDifficulty, modData.mods, modData.help);
-    game.burningwheel.modifiers.activateSocketListeners();
-    game.burningwheel.modifiers.render(true);
-
+    await migrateData();
+    await dialogs.initializeRollPanels();
     RegisterMacros();
-});
-
-Hooks.on("renderSidebarTab", async (_data, html: JQuery) => {
-    if (html.prop("id") === "combat") { // this is the combat tab
-        DuelOfWitsDialog.addSidebarControl(html);
-        FightDialog.addSidebarControl(html);
-        RangeAndCoverDialog.addSidebarControl(html);
-    }
 });
 
 function registerHelpers() {
